cosmetic changes; move torch_gradient to util (#27)
